- WIP: Updated Obsidian view models to use legacy enum types (instead of numbers).

+++ Rock.Blocks/Event/RegistrationEntry.cs
@@ -3159,7 +3159,7 @@ namespace Rock.Blocks.Event
                 FamilyMembers = familyMembers,
                 MaxRegistrants = context.RegistrationSettings.MaxRegistrants ?? 25,
                 ShowSmsOptIn = context.RegistrationSettings.ShowSmsOptIn,
-                RegistrantsSameFamily = ( int ) context.RegistrationSettings.RegistrantsSameFamily,
+                RegistrantsSameFamily = context.RegistrationSettings.RegistrantsSameFamily,
                 ForceEmailUpdate = forceEmailUpdate,
                 RegistrarOption = context.RegistrationSettings.RegistrarOption,
                 Cost = baseCost,
